requirements.txt
azure-functions==1.21.3
defusedxml==0.7.1
msal==1.31.1
requests==2.32.3
